+++ src/main/java/com/kyouseipro/neo/recycle/item/repository/RecycleItemParameterBinder.java
@@ -27,20 +27,20 @@ public class RecycleItemParameterBinder {
         return index;
     }
 
-    public static int bindDeleteForIds(PreparedStatement ps, List<Integer> ids) throws SQLException {
+    public static int bindDeleteForIds(PreparedStatement ps, List<Long> ids) throws SQLException {
         int index = 1;
         ps.setInt(index++, Enums.state.DELETE.getCode());
-        for (Integer id : ids) {
-            ps.setInt(index++, id);
+        for (Long id : ids) {
+            ps.setLong(index++, id);
         }
         ps.setInt(index++, Enums.state.DELETE.getCode());
         return index;
     }
 
-    public static int bindDownloadCsvForIds(PreparedStatement ps, List<Integer> ids) throws SQLException {
+    public static int bindDownloadCsvForIds(PreparedStatement ps, List<Long> ids) throws SQLException {
         int index = 1;
-        for (Integer id : ids) {
-            ps.setInt(index++, id);
+        for (Long id : ids) {
+            ps.setLong(index++, id);
         }
         ps.setInt(index++, Enums.state.DELETE.getCode());
         return index;
