+++ src/main/java/com/kyouseipro/neo/recycle/maker/repository/RecycleMakerParameterBinder.java
@@ -32,20 +32,20 @@ public class RecycleMakerParameterBinder {
         return index;
     }
 
-    public static int bindDeleteByIds(PreparedStatement ps, List<Integer> ids) throws SQLException {
+    public static int bindDeleteByIds(PreparedStatement ps, List<Long> ids) throws SQLException {
         int index = 1;
         ps.setInt(index++, Enums.state.DELETE.getCode());
-        for (Integer id : ids) {
-            ps.setInt(index++, id);
+        for (Long id : ids) {
+            ps.setLong(index++, id);
         }
         ps.setInt(index++, Enums.state.DELETE.getCode());
         return index;
     }
 
-    public static int bindDownloadCsvByIds(PreparedStatement ps, List<Integer> ids) throws SQLException {
+    public static int bindDownloadCsvByIds(PreparedStatement ps, List<Long> ids) throws SQLException {
         int index = 1;
-        for (Integer id : ids) {
-            ps.setInt(index++, id);
+        for (Long id : ids) {
+            ps.setLong(index++, id);
         }
         ps.setInt(index++, Enums.state.DELETE.getCode());
         return index;
